Backend/requirements.txt
arxiv==1.4.3
PyMuPDF==1.22.3
requests==2.31.0
tiktoken==0.4.0
tenacity==8.2.2
pybase64==1.2.3
Pillow==9.4.0
openai==0.27.8
markdown
jieba
gradio==3.20.1
flask
flask-cors==3.0.10
bs4==0.0.1
urllib3==1.25.11
sphinx-markdown-tables
sphinxemoji
recommonmark
sphinx-book-theme
fastapi
uvicorn
pydantic
sqlalchemy
python-jose